Shoppingmall/src/main/java/com/shoppingmall/service/UserService.java
package com.shoppingmall.service;

import com.shoppingmall.entity.User;
import com.shoppingmall.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional  // 문제 발생시 롤백 시켜주는 어노테이션
@RequiredArgsConstructor    // 생성자 생성
public class UserService {

    private final UserRepository userRepository;

    public User saveUser(User user){
        validateDuplicateUser(user);
        return userRepository.save(user);
    }

    private void validateDuplicateUser(User user) {
        User findUser = userRepository.findByEmail(user.getEmail());
        if(findUser != null){
            throw new IllegalStateException("이미 가입된 회원");
        }
    }
}
